requirements.txt
flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.2.0
Flask-MySQLdb==1.0.1 